# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Rental Agency API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  agency: "agency",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  agencyName?: string | null;
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface RegisterCustomerInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface RegisterAgencyInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  agencyName: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Car {
  id: number;
  vehicleModel: string;
  vehicleNumber: string;
  seatingCapacity: number;
  rentPerDay: number;
  agencyId: number;
  agencyName: string;
  available: boolean;
  createdAt: string;
}

export interface CarInput {
  /** @minLength 2 */
  vehicleModel: string;
  /** @minLength 2 */
  vehicleNumber: string;
  /** @minimum 1 */
  seatingCapacity: number;
  /** @minimum 0 */
  rentPerDay: number;
}

export interface Booking {
  id: number;
  carId: number;
  customerId: number;
  startDate: string;
  numberOfDays: number;
  totalCost: number;
  createdAt: string;
}

export interface BookingInput {
  carId: number;
  startDate: string;
  /**
   * @minimum 1
   * @maximum 30
   */
  numberOfDays: number;
}

export interface BookingWithDetails {
  id: number;
  carId: number;
  vehicleModel: string;
  vehicleNumber: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  startDate: string;
  numberOfDays: number;
  totalCost: number;
  createdAt: string;
}