# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Car Rental Agency API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: cars
    description: Car management
  - name: bookings
    description: Booking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register/customer:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCustomerInput"
      responses:
        "201":
          description: Customer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register/agency:
    post:
      operationId: registerAgency
      tags: [auth]
      summary: Register a new car rental agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAgencyInput"
      responses:
        "201":
          description: Agency registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login for both customers and agencies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List all available cars (public)
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
    post:
      operationId: addCar
      tags: [cars]
      summary: Add a new car (agency only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarInput"
      responses:
        "201":
          description: Car added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /cars/{id}:
    put:
      operationId: updateCar
      tags: [cars]
      summary: Update a car (agency only, own cars)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarInput"
      responses:
        "200":
          description: Car updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Car not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a car (customer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not allowed (agencies cannot book)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agency/cars:
    get:
      operationId: getAgencyCars
      tags: [cars]
      summary: Get cars added by the logged-in agency
      responses:
        "200":
          description: Agency's cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /agency/bookings:
    get:
      operationId: getAgencyBookings
      tags: [bookings]
      summary: Get all bookings for the agency's cars (agency only)
      responses:
        "200":
          description: List of bookings with customer info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingWithDetails"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, agency]
        agencyName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    RegisterCustomerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
      required:
        - name
        - email
        - password

    RegisterAgencyInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        agencyName:
          type: string
          minLength: 2
        phone:
          type: string
      required:
        - name
        - email
        - password
        - agencyName

    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    Car:
      type: object
      properties:
        id:
          type: integer
        vehicleModel:
          type: string
        vehicleNumber:
          type: string
        seatingCapacity:
          type: integer
        rentPerDay:
          type: number
        agencyId:
          type: integer
        agencyName:
          type: string
        available:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - vehicleModel
        - vehicleNumber
        - seatingCapacity
        - rentPerDay
        - agencyId
        - agencyName
        - available
        - createdAt

    CarInput:
      type: object
      properties:
        vehicleModel:
          type: string
          minLength: 2
        vehicleNumber:
          type: string
          minLength: 2
        seatingCapacity:
          type: integer
          minimum: 1
        rentPerDay:
          type: number
          minimum: 0
      required:
        - vehicleModel
        - vehicleNumber
        - seatingCapacity
        - rentPerDay

    Booking:
      type: object
      properties:
        id:
          type: integer
        carId:
          type: integer
        customerId:
          type: integer
        startDate:
          type: string
        numberOfDays:
          type: integer
        totalCost:
          type: number
        createdAt:
          type: string
      required:
        - id
        - carId
        - customerId
        - startDate
        - numberOfDays
        - totalCost
        - createdAt

    BookingInput:
      type: object
      properties:
        carId:
          type: integer
        startDate:
          type: string
        numberOfDays:
          type: integer
          minimum: 1
          maximum: 30
      required:
        - carId
        - startDate
        - numberOfDays

    BookingWithDetails:
      type: object
      properties:
        id:
          type: integer
        carId:
          type: integer
        vehicleModel:
          type: string
        vehicleNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        startDate:
          type: string
        numberOfDays:
          type: integer
        totalCost:
          type: number
        createdAt:
          type: string
      required:
        - id
        - carId
        - vehicleModel
        - vehicleNumber
        - customerName
        - customerEmail
        - startDate
        - numberOfDays
        - totalCost
        - createdAt