# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Rental Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  carId: number;
  customerId: number;
  startDate: string;
  numberOfDays: number;
  totalCost: number;
  createdAt: string;
}