# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Rental Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  carId: number;
  startDate: string;
  /**
   * @minimum 1
   * @maximum 30
   */
  numberOfDays: number;
}