# lib/api-zod/src/generated/types/bookingWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Rental Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingWithDetails {
  id: number;
  carId: number;
  vehicleModel: string;
  vehicleNumber: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  startDate: string;
  numberOfDays: number;
  totalCost: number;
  createdAt: string;
}