# lib/api-zod/src/generated/types/car.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Rental Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface Car {
  id: number;
  vehicleModel: string;
  vehicleNumber: string;
  seatingCapacity: number;
  rentPerDay: number;
  agencyId: number;
  agencyName: string;
  available: boolean;
  createdAt: string;
}