# lib/api-zod/src/generated/types/carInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Rental Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface CarInput {
  /** @minLength 2 */
  vehicleModel: string;
  /** @minLength 2 */
  vehicleNumber: string;
  /** @minimum 1 */
  seatingCapacity: number;
  /** @minimum 0 */
  rentPerDay: number;
}