# lib/api-zod/src/generated/types/registerAgencyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Rental Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterAgencyInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  agencyName: string;
  phone?: string;
}